# weeks/week-00/code/openapi.yaml
openapi: 3.0.3
info:
  title: SOA Course Service API (Week 00)
  version: 1.0.0
  description: >-
    Hợp đồng dịch vụ mẫu cho Tuần 00 học phần Kiến trúc hướng dịch vụ (SOA) - VNU-UET.
    Minh họa mẫu thiết kế Resource Naming (JJ Geewax) và tư duy API-as-a-Product (Bruno Pedro).
  contact:
    name: Bộ môn Công nghệ Phần mềm - VNU-UET
    email: fit@vnu.edu.vn
servers:
  - url: http://localhost:3000
    description: Môi trường thử nghiệm cục bộ (Local Development Server)

paths:
  /health:
    get:
      summary: Kiểm tra tình trạng sức khỏe dịch vụ (Health Check)
      description: Trả về trạng thái hoạt động của Service Provider và tình trạng kết nối tới Database.
      responses:
        '200':
          description: Dịch vụ đang hoạt động ổn định
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: UP
                  database:
                    type: string
                    example: In-Memory Mock Fallback
                  timestamp:
                    type: string
                    format: date-time
                    example: '2026-09-04T15:00:00.000Z'

  /api/v1/courses:
    get:
      summary: Liệt kê danh sách học phần (List Pattern)
      description: Trả về toàn bộ danh sách các tài nguyên học phần thuộc trường đại học.
      responses:
        '200':
          description: Lấy danh sách thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResponse'
        '500':
          description: Lỗi máy chủ nội bộ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Tạo mới một học phần (Create Pattern)
      description: Tiếp nhận thông tin học phần mới, thực hiện xác thực và trả về mã 201 Created kèm vị trí tài nguyên.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseInput'
      responses:
        '201':
          description: Tạo học phần thành công
          headers:
            Location:
              description: Đường dẫn URI dẫn tới tài nguyên vừa tạo
              schema:
                type: string
                example: /api/v1/courses/65e5f1a0b123456789abcdef
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Dữ liệu gửi lên thiếu hoặc không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Trùng lặp mã học phần (Conflict)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/courses/{id}:
    get:
      summary: Lấy chi tiết một học phần (Get Pattern)
      description: Truy xuất thông tin chi tiết của một tài nguyên học phần dựa trên ID định danh.
      parameters:
        - name: id
          in: path
          required: true
          description: ID định danh duy nhất của học phần
          schema:
            type: string
            example: 65e5f1a0b123456789abcdef
      responses:
        '200':
          description: Tìm thấy thông tin học phần
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Định dạng ID không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Không tìm thấy học phần tương ứng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Xóa một học phần (Delete Pattern)
      description: Xóa tài nguyên học phần ra khỏi hệ thống. Trả về mã 204 No Content khi hoàn tất.
      parameters:
        - name: id
          in: path
          required: true
          description: ID định danh của học phần cần xóa
          schema:
            type: string
            example: 65e5f1a0b123456789abcdef
      responses:
        '204':
          description: Xóa thành công (Không có nội dung trả về)
        '400':
          description: Định dạng ID không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Không tìm thấy học phần để xóa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CourseInput:
      type: object
      required:
        - code
        - title
        - credits
      properties:
        code:
          type: string
          description: Mã định danh duy nhất của môn học (Viết hoa)
          example: INT3111
        title:
          type: string
          description: Tên đầy đủ của môn học
          example: Kiến trúc hướng dịch vụ (SOA)
        credits:
          type: integer
          minimum: 1
          maximum: 10
          description: Số lượng tín chỉ tích lũy
          example: 3
        department:
          type: string
          description: Đơn vị phụ trách giảng dạy
          example: Bộ môn Công nghệ Phần mềm

    Course:
      allOf:
        - type: object
          properties:
            _id:
              type: string
              description: ID tự sinh của cơ sở dữ liệu
              example: 65e5f1a0b123456789abcdef
            createdAt:
              type: string
              format: date-time
              example: '2026-09-04T12:00:00.000Z'
            updatedAt:
              type: string
              format: date-time
              example: '2026-09-04T12:00:00.000Z'
        - $ref: '#/components/schemas/CourseInput'

    CourseListResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: '#/components/schemas/Course'
        total:
          type: integer
          description: Tổng số lượng học phần hiện có
          example: 2

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Dữ liệu không hợp lệ
        message:
          type: string
          example: Trường code, title và credits là bắt buộc.
        details:
          type: string
          example: Ràng buộc kiểm tra Schema thất bại